Move reduce operators into reduce_executor.rs
